% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backloggd-api.R
\name{filters}
\alias{filters}
\title{Setting the filters for the search}
\usage{
filters(type = "", release_year = "", genre = "", game_status = "")
}
\arguments{
\item{type}{'Type' filter.}

\item{release_year}{'Release Year' filter.}

\item{genre}{'Genre' filter.}

\item{game_status}{'Played Status' filter.}
}
\value{
A string of the form \verb{field1:value1;field2:value2...}.
}
\description{
Generate the string containing the filters used to retrieve data from
backloggd.com. Available filters are 'Type', 'Release Year', 'Genre' and
'Played Status'. Refer to the website for the possible values of these filters.
}
