% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backloggd-api.R
\name{getdata}
\alias{getdata}
\title{Getting the data for a user}
\usage{
getdata(user, filters = "")
}
\arguments{
\item{user}{Name of the user to retrieve the library from.}

\item{filters}{Optional list of filters.}
}
\value{
A list where \verb{$filters} shows the applied filters and their value and \verb{$df} is a data frame with the IDs and names of the games.
}
\description{
Retrieve the library of a user according to a specified set of filters.
}
